API

MINDFulCompanion.getmultilayernodegroups — Method

getmultilayernodegroups(mlg) -> Vector{Vector{Int64}}

Return multilayer nodes in Vector{Vector{Int}} format. Used for plotting

MINDFulCompanion.jointrmsagenerilizeddijkstra! — Method

jointrmsagenerilizeddijkstra!(
    ibn::MINDFul.IBN,
    idagnode::MINDFul.IntentDAGNode{R<:MINDFul.DomainConnectivityIntent};
    optimizepaths,
    minrate,
    bordergroomingenabled,
    time
)

MINDFulCompanion.jointrmsagenerilizeddijkstra! — Method

jointrmsagenerilizeddijkstra!(
    myibn::MINDFul.IBN,
    neibn::MINDFul.IBN,
    idagnode::MINDFul.IntentDAGNode{T<:MINDFul.ConnectivityIntent},
    iid::MINDFul.InterIntent{R<:MINDFul.IntentDirection};
    args...
)

Handles interdomain connectivity intents

MINDFulCompanion.mlnodegraphtomlgraph — Method

mlnodegraphtomlgraph(
    ibn::MINDFul.IBN,
    rate::Real;
    bordergroomingenabled
) -> NestedGraphs.NestedGraph{Int64, AttributeGraphs.OAttributeGraph{Int64, WrappedMultiGraphs.MultiDiGraph{Int64}, Tuple{MINDFulCompanion.NodeType, Int64}, MINDFulCompanion.LinkCostVector, Missing}, AttributeGraphs.OAttributeGraph{Int64, WrappedMultiGraphs.MultiDiGraph{Int64}, Tuple{MINDFulCompanion.NodeType, Int64}, MINDFulCompanion.LinkCostVector, Missing}}

Add virtual links and ip cost

MINDFulCompanion.mlnodegraphtomlgraph — Method

mlnodegraphtomlgraph(
    ngr::NestedGraphs.NestedGraph
) -> NestedGraphs.NestedGraph{Int64, AttributeGraphs.OAttributeGraph{Int64, WrappedMultiGraphs.MultiDiGraph{Int64}, Tuple{MINDFulCompanion.NodeType, Int64}, MINDFulCompanion.LinkCostVector, Missing}, AttributeGraphs.OAttributeGraph{Int64, WrappedMultiGraphs.MultiDiGraph{Int64}, Tuple{MINDFulCompanion.NodeType, Int64}, MINDFulCompanion.LinkCostVector, Missing}}

Convert a nested graph composed by MLNodes to a multi layer nested graph. New multi layer graph loses all info about nested stuctures Return also a data structure to nodes of the new nested graph to the nodes of the old one.

There are 2 layers in the multi layer graph: PHY, IP

Index

MINDFulCompanion.getmultilayernodegroups

MINDFulCompanion.jointrmsagenerilizeddijkstra!

MINDFulCompanion.jointrmsagenerilizeddijkstra!

MINDFulCompanion.mlnodegraphtomlgraph

MINDFulCompanion.mlnodegraphtomlgraph